pyproject.toml
[project]
name = "data-explorer"
version = "1.0.0"
description = "Data explorer web app proof-of-concept using Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.46.0",
]
